Convert WordPressData from Swift package to Xcode target
To much Objective-C code using logic implemented elsewhere in Swift for
us to be able to move the files in two separate Objective-C- and Swift-
only packages.

Refer to https://github.com/wordpress-mobile/WordPress-iOS/pull/24166 as
an experiment for this that shows the various build failure that even a
partial attempt at moving some files results in.