Convert a number of interconnected `Blog` methods to Swift
As part of #24165, it's convenient to re-implement part of the
Objective-C models to Swift, to remove code in Objective-C that accesses
logic defined in the Swift layer.

Regardless of where WordPressData will end up, this kind of rewrites are
advantageous because they bring us a few inches closer to our goal of
no longer having Objective-C code, with all the implementation benefits
that will produce.