Inject `ApiCredentials` instead of reading them internally (#24435)
* Inject `ApiCredentials` `.client` and `.secret`

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` Google client details

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` Tenor API key

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` Sentry DSN

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` `encryptedLogKey`

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` `docsBotId`

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.