Decouple Keystone from `ApiCredentials` (#24439)
* Inject `ApiCredentials` `.client` and `.secret`

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` Google client details

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` Tenor API key

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` Sentry DSN

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` `encryptedLogKey`

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Inject `ApiCredentials` `docsBotId`

That's instead of doing a direct read under the hood.

The aim of this is to get to a place where Keystone is agnostic of
`ApiCredentials`.

* Define secrets type in BuildSettingsKit

* Add secrets property to `BuildSettings`

* Configure secrets in `BuildSettings` using `ApiCredentials`

* Replace accessing secrets from `ApiCredentials` with `BuildSettings`

* Remove `Secrets.swift` from Keystone, add to Reader

* Add placeholder in script to generate credentials for Reader

* Do not upload encrypted logs during unit tests

After migrating to reading secrets from `BuildSecrets`, unit tests
started crashing in the background thread attempting the encrypted logs
upload.

This is possibly related to the changes in how the secrets are accessed,
but anyways we should not upload logs from the unit tests as that's just
noise.

* Add demo pipeline that builds Reader for ASC distribution (#24440)