Add dedicated secrets file for Reader (#24463)
* Add dedicated secrets file for Reader

As a result, the secrets files for the other apps updated too.
That's because since the last update we removed the App Center API key
from them.

See also
https://github.com/wordpress-mobile/WordPress-iOS/commit/9348997a1dbed073c541ee7ce31bf19fae0d7e06

* Set `WP_APP_URL_SCHEME` for Reader

* Use appropriate `WP_BUILD_CONFIGURATION` for Reader debug and release

* Make Reader use its own `ApiCredentials` value

* Fix `Bundle` for Notifications in Reader Tab VC

We need to use the bundle to the class, otherwise UIKit will look for
the storyboard in the app (Reader) when the file is in Keystone.

* Constrain `NibLoadable` to `AnyObject`

This is so we can make it use the bundle of the class conforming to it
instead of using `.main`, which fails if the class and its nib are in a
framework (Keystone)