Use the same `key` for the Reader build step across pipelines (#24488)
See how
https://buildkite.com/automattic/wordpress-ios/builds/27417/steps?jid=01964b59-300d-46d2-bf40-33558caa03fa
failed because the pipeline used `reader_asc_build` instead of
`build_asc_reader`.

I decided to consolidate to `build_asc_reader` because of the
information hierarchy: build > build for ASC > build for ASC, the Reader
app.

When we'll get to split build from upload for the other apps, we'll be
able to have auto-complete friendly keys, su as `build_asc_wordpress` or
`build_firebase_jetpack`.