Reset Modules as an Xcode local package with tests in CI (#24518)
* Add a dummy local package—We'll try to put the Modules there next

* Replace ModulesAlterantive with Modules

* Re-enable all the test targets in WordPressUnitTests

* Delete unnecessary OCLint scheme

Given OCLint is something we should drop soon, because it's only for
Objective-C and we want to reduce that code surface.

* Configure WordPressFlux scheme to run tests

* Move Modules package reference from workspace to project

* Use the shared setup CI script in the complete code freeze step

* Add workaround for package path support missing

* Point `color-studio` to a commit instead of a branch

Locking dependencies to commits makes updating them a bit more
burdensome, but at least it makes the update intentional and not
automatic and surprising.

This change was done to test Xcode's behavior in monitoring
`Modules/Package.swift` and updating `Modules/Package.resolved`.

See
https://github.com/wordpress-mobile/WordPress-iOS/pull/24518#discussion_r2079810655